test(Run Info): Move some tests from Student to User test file
#229
